Update Program.cs
minor edit

+++ Fib/Program.cs
@@ -1,18 +1,19 @@
 ﻿using System;
+using System.Numerics;
 namespace Hello
 {
     class Fibonacci
     {
         static void Main(string[] args)
         {
-            int a, b;
+            BigInteger a, b;
             a = 0;
             b = 1;
             Console.Write("Enter sequence length: ");
             int n = int.Parse(Console.ReadLine());
             for (int i = 0; i <= n; i++)
             {
-                int fib = a;
+                BigInteger fib = a;
                 a = b;
                 b = fib + a;
                 Console.WriteLine("Fibonacci sequence : "+fib);
